Implement support for fabric client again
